#SORQA-961 : created test and refactored code
