+++ sormas-e2e-tests/src/main/java/org/sormas/e2etests/pages/application/aCommonComponents/SideCards.java
@@ -1,0 +1,49 @@
+/*
+ * SORMAS® - Surveillance Outbreak Response Management & Analysis System
+ * Copyright © 2016-2022 Helmholtz-Zentrum für Infektionsforschung GmbH (HZI)
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package org.sormas.e2etests.pages.application.aCommonComponents;
+
+import org.openqa.selenium.By;
+
+public class SideCards {
+
+  public static final By ADDED_SAMPLES_IN_SAMPLE_CARD =
+      By.xpath("//*[@location='samples']//*[@class='v-slot v-slot-s-list']/div/div/div/div");
+  public static final By EDIT_SAMPLE_BUTTON = By.xpath("//div[contains(@id, 'edit-sample')]");
+
+  public static By checkTextInHandoverSideComponent(String text) {
+    return By.xpath(String.format("//div[@location='sormasToSormas']//div[text()='%s']", text));
+  }
+
+  public static By checkTextInSampleSideComponent(String text) {
+    return By.xpath(
+        String.format("//div[contains(@location,'samples')]//div[contains(text(), '%s')]", text));
+  }
+
+  public static By checkTextInImmunizationSideComponent(String text) {
+    return By.xpath(
+        String.format(
+            "//div[contains(@location,'vaccinations')]//div[contains(text(), '%s')]", text));
+  }
+
+  public static By checkTextInReportSideComponent(String text) {
+    return By.xpath(
+        String.format(
+            "//div[text()='Meldevorg\u00E4nge']/../../../../../..//div[text()='%s']", text));
+  }
+}
